receiver-service/requirements.txt
fastapi==0.115.0
google-cloud-pubsub==2.23.0
pydantic==2.9.2
uvicorn[standard]==0.30.6
